requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
img2pdf>=0.5.0
python-pptx>=0.6.21
